Adds gameOver state & hoverTile to createNonogram
